requirements.txt
mcp>=0.9.0
jira>=3.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
git+https://github.com/cmchase/team-reports.git 